+++ src/tightdb/util/thread.cpp
@@ -1,3 +1,23 @@
+/*************************************************************************
+ *
+ * TIGHTDB CONFIDENTIAL
+ * __________________
+ *
+ *  [2011] - [2012] TightDB Inc
+ *  All Rights Reserved.
+ *
+ * NOTICE:  All information contained herein is, and remains
+ * the property of TightDB Incorporated and its suppliers,
+ * if any.  The intellectual and technical concepts contained
+ * herein are proprietary to TightDB Incorporated
+ * and its suppliers and may be covered by U.S. and Foreign Patents,
+ * patents in process, and are protected by trade secret or copyright law.
+ * Dissemination of this information or reproduction of this material
+ * is strictly forbidden unless prior written permission is obtained
+ * from TightDB Incorporated.
+ *
+ **************************************************************************/
+
 #include <stdexcept>
 
 #include <tightdb/util/thread.hpp>
@@ -201,151 +221,6 @@ void RobustMutex::mark_as_consistent() TIGHTDB_NOEXCEPT
 #endif
 }
 
-string PlatformSpecificCondVar::internal_naming_prefix = "/RealmsBigFriendlySemaphore";
-
-void PlatformSpecificCondVar::set_resource_naming_prefix(std::string prefix)
-{
-    internal_naming_prefix = prefix + "RLM";
-}
-
-PlatformSpecificCondVar::PlatformSpecificCondVar()
-{
-#ifdef TIGHTDB_HAVE_PTHREAD_PROCESS_SHARED
-    m_shared_part = 0;
-    m_sem = 0;
-#else // !TIGHTDB_HAVE_PTHREAD_PROCESS_SHARED
-    throw runtime_error("No support for process-shared condition variables");
-#endif
-}
-
-
-
-
-
-
-void PlatformSpecificCondVar::close() TIGHTDB_NOEXCEPT
-{
-    if (m_sem) { // true if emulating a process shared condvar
-        sem_close(m_sem);
-        m_sem = 0;
-        return; // we don't need to clean up the SharedPart
-    }
-    // we don't do anything to the shared part, other CondVars may shared it
-    m_shared_part = 0;
-}
-
-
-PlatformSpecificCondVar::~PlatformSpecificCondVar() TIGHTDB_NOEXCEPT
-{
-    close();
-}
-
-
-
-void PlatformSpecificCondVar::set_shared_part(SharedPart& shared_part, std::string path, std::size_t offset_of_condvar)
-{
-    TIGHTDB_ASSERT(m_shared_part == 0);
-    close();
-    m_shared_part = &shared_part;
-    static_cast<void>(path);
-    static_cast<void>(offset_of_condvar);
-#ifdef TIGHTDB_CONDVAR_EMULATION
-    m_sem = get_semaphore(path,offset_of_condvar);
-#endif
-}
-
-sem_t* PlatformSpecificCondVar::get_semaphore(std::string path, std::size_t offset)
-{
-#if 0
-    struct stat su;
-    // if stat does not work, we default to 0/0 for device and inode.
-    su.st_dev = 0; su.st_ino = 0;
-    stat(path.c_str(), &su);
-    dev_t device = su.st_dev;
-    ino_t inode = su.st_ino;
-    uint64_t magic = (device+1);
-    magic *= (inode+1);
-#else
-    uint64_t magic = 1;
-    for (unsigned int i=0; i<path.length(); ++i)
-        magic *= 1 + (unsigned int)(path[i]);
-#endif
-    std::string name = internal_naming_prefix;
-    magic *= (offset+1);
-    name += 'A'+(magic % 23);
-    magic /= 23;
-    name += 'A'+(magic % 23);
-    magic /= 23;
-    name += 'A'+(magic % 23);
-    magic /= 23;
-    TIGHTDB_ASSERT(m_shared_part);
-    if (m_sem == 0) {
-        m_sem = sem_open(name.c_str(), O_CREAT, S_IRWXG | S_IRWXU, 0);
-        // FIXME: error checking
-    }
-    return m_sem;
-}
-
-
-void PlatformSpecificCondVar::init_shared_part(SharedPart& shared_part) {
-#ifdef TIGHTDB_CONDVAR_EMULATION
-    shared_part.waiters = 0;
-    shared_part.signal_counter = 0;
-#else
-    pthread_condattr_t attr;
-    int r = pthread_condattr_init(&attr);
-    if (TIGHTDB_UNLIKELY(r != 0))
-        attr_init_failed(r);
-    r = pthread_condattr_setpshared(&attr, PTHREAD_PROCESS_SHARED);
-    TIGHTDB_ASSERT(r == 0);
-    r = pthread_cond_init(&shared_part.m_impl, &attr);
-    int r2 = pthread_condattr_destroy(&attr);
-    TIGHTDB_ASSERT(r2 == 0);
-    static_cast<void>(r2);
-    if (TIGHTDB_UNLIKELY(r != 0))
-        init_failed(r);
-#endif // TIGHTDB_CONDVAR_EMULATION
-}
-
-TIGHTDB_NORETURN void PlatformSpecificCondVar::init_failed(int err)
-{
-    switch (err) {
-        case ENOMEM:
-            throw bad_alloc();
-        default:
-            throw runtime_error("pthread_cond_init() failed");
-    }
-}
-
-void PlatformSpecificCondVar::handle_wait_error(int err)
-{
-#ifdef TIGHTDB_HAVE_ROBUST_PTHREAD_MUTEX
-    if (err == ENOTRECOVERABLE)
-        throw RobustMutex::NotRecoverable();
-    if (err == EOWNERDEAD)
-        return;
-#else
-    static_cast<void>(err);
-#endif
-    TIGHTDB_TERMINATE("pthread_mutex_lock() failed");
-}
-
-TIGHTDB_NORETURN void PlatformSpecificCondVar::attr_init_failed(int err)
-{
-    switch (err) {
-        case ENOMEM:
-            throw bad_alloc();
-        default:
-            throw runtime_error("pthread_condattr_init() failed");
-    }
-}
-
-TIGHTDB_NORETURN void PlatformSpecificCondVar::destroy_failed(int err) TIGHTDB_NOEXCEPT
-{
-    if (err == EBUSY)
-        TIGHTDB_TERMINATE("Destruction of condition variable in use");
-    TIGHTDB_TERMINATE("pthread_cond_destroy() failed");
-}
 
 
 CondVar::CondVar(process_shared_tag)